Update RDF function
